Improve SSLproxy parser and Utmfw preprocessor for snort
ATTENTION: It seems like the Utmfw preprocessor can modify the src and
dst addresses only for the packet which contains the SSLproxy line (the
first packet in the connection), not for all the packets in the stream.
We should find a way to apply the same src and dst addresses to all the
packets in the stream, perhaps by patching the stream preprocessor with
the code in Utmfw preprocessor.