Modify SessionControlBlock addresses with SSLproxy line addresses and use it to translate all packet addresses in connection
"preprocessor normalize_tcp: block" should be disabled for this
preprocessor to work properly, i.e. do not allow packet drops during TCP
normalization. Otherwise, connections fail and snort keeps reporting the
original/unmodified packets. Other normalize_tcp options seem to be
fine.

Snort sometimes fails to feed the packets from the server to the client
to this preprocessor after we translate the SessionControlBlock
addresses, as a result we cannot translate the addresses in those
packets. This random issue seems to happen because the session cache
functions get confused after we modify the SessionControlBlock of the
session. Apparently, we are not supposed to modify the
SessionControlBlock here. Unfortunately, this issue seems to happen in
about 50% of the connections.

If this preprocessor successfully translates the addresses, the rules
make use of and the logs report those translated addresses correctly.

Note that we cannot translate the addresses in the packets until we
receive the first packet with the SSLproxy line. Those are usually the
three-way handshake packets.

TODO: If we cannot solve the issue mentioned above, we should patch the
Session preprocessor with this code, so that it works all the time.