Fix address translation for packets to client, and reenable preprocessor normalize_tcp: block
The issue was that snort function to detect packet direction was
comparing IP addresses only, but all inline packets to snort here have
the loopback address. So we fix this by also comparing port numbers
along with IP addresses. This also fixes the blocked packets with tcp
normalization.
TODO: Check if this breaks anything with the IDS mode. (This fix is for
the inline IPS mode.)