Fix rules with flow established flag
Setting the packet established flag by scb->ha_state.session_flags |=
SSNFLAG_ESTABLISHED fixes the rules with flow: established option.
However, a copy of the packet with the SSLproxy line still remains as
stale, and eventually times out, or is processed while snort terminates,
whichever happens first. But fortunately this issue does not trigger the
alert associated with that rule for a second time.