Set login shells of admin and user users to ctlr
Update related code and documentation accordingly.

So admin and user users can log in to the system, but cannot drop to a
shell command line now. They can execute only the controller commands
via the ctlr script in their home folders, as it is their login shell.

This change requires modification in the Android and Windows user
interface code too.