Update snort to 2.9.16, rename utmfw preproc to sslproxy preproc
Update config and ports accordingly.
But keep 2.9.15 in the repo.

Note that daq 2.0.7 is for windows support only: "Bump to 2.0.7 and
added DAQ project file for supporting windows 64-bit." There is no
source code changes other than a couple of lines for windows project
file support. Plus, there are issues with its configure, which still
says it is for daq 2.0.6 anyway. So we still use daq 2.0.6.