Do not process log files larger than $MaxLogFileSize
We should not pass large log files to stats and logs functions, or to
grep, tail, head, cat, and similar programs. Otherwise, they run for
minutes causing very heavy cpu, memory, and disk load.

$MaxLogFileSize in MBs can be configured on the system WUI conf page.

NOTE: We perhaps need a stats collection daemon for our stats charts,
something similar to collectd. But it is not enough, because our stats
charts and top lists display strings too, not just numbers fetched from
rrd files. Plus the problem is not just about stats charts, but also
logs functions too, which can not make use of such a daemon. So, a stats
daemon is not enough.

Also, note that this solution was able to successfully handle DNS floods
caused by Firefox, where the named.log file had reached 1GB in size, and
pf states reached 20K. Can a custom stats daemon handle that?

Therefore, the solution in this commit is perhaps not just a workaround
but the right solution.