Decrease default $MaxLogFileSize to 10 MB
The default file size for log rotation is 10 MB for most log files in
newsyslog.conf.

This change was necessary because named.log file of 1GB in size was
compressed to a 12 MB .gz file, which we should not try to uncompress
and process on the stats and logs pages.