Add SSLproxy rule editor, SPRE, similar to PFRE
We now support the proxyspecs, filtering rules, and configuration
options of the next version of SSLproxy, to be released soon, on a rule
editor very similar to PFRE. The SPRE code is based on the PFRE code.

- Structured proxyspecs are entered on an edit box, for now. It must be
possible to configure structured proxyspecs on a new SPRE page when the
user clicks its edit button.
- Include files can only be selected, but can be configured by loading
separately to SPRE for now. It must be possible to configure include
files on a new SPRE page when the user clicks its edit button.

Also,
- Fix config upload in PFRE by fixing tmp file path
- Convert edit links to buttons
- Move edit links from the tail to the head
- Move common functions of PFRE and SPRE to the Model
- Clean up now unused code, e.g. sslproxy module does not use the
conf.php under lib now